Revert "Merge High Availabillity features into blockchain-service  (#1424)"
This reverts commit 6cd5c62030cae996d16a876916ae2d9c0f681573.

+++ models/reputation_data.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
         dh_identity: DataTypes.STRING,
         offer_id: DataTypes.STRING,
         reputation_delta: DataTypes.STRING,
-        timestamp: DataTypes.BIGINT,
+        timestamp: DataTypes.INTEGER,
     }, {
         tableName: 'reputation_data',
     });
